+++ API/ICar.API/ViewModels/Car/CarViewModel.cs
@@ -12,7 +12,7 @@ namespace ICar.API.ViewModels.Car
         public int MakeDate { get; set; }
         public int MakedDate { get; set; }
         public double KilometersTraveled { get; set; }
-        public double Price { get; set; }
+        public int Price { get; set; }
         public bool AcceptsChange { get; set; }
         public bool IpvaIsPaid { get; set; }
         public bool IsLicensed { get; set; }
